+++ MLEM.Extended/Font/GenericStashFont.cs
@@ -53,8 +53,8 @@ namespace MLEM.Extended.Font {
         }
 
         /// <inheritdoc />
-        protected override Vector2 MeasureChar(char c) {
-            return this.Font.MeasureString(c.ToCachedString());
+        protected override float MeasureChar(char c) {
+            return this.Font.MeasureString(c.ToCachedString()).X;
         }
 
         private static float CalculateLineHeight(SpriteFontBase font) {